pyproject.toml
[project]
name = "sbfa"
version = "0.1.0"
description = "SilverBulletForAI - Multi-AI Team + RAG + MCP System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "surrealdb>=0.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.43.0",
    "google-genai>=1.0.0",
    "openai>=1.60.0",
    "python-dotenv>=1.0.0",
    "pyjwt[crypto]>=2.9.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sbfa"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
